Adjust CAS & multi-register models to not ignore nil values
While doing some fiddling with Knossos as a standalone library,
outside of Jepsen, I found something odd when analyzing the
following trace.

    [
    ;; Set register to something

    {:time   30 :type :invoke :f :write :value  132 :process 0}
    {:time   40 :type :ok     :f :write :value  132 :process 0}

    ;; This succeeds, as expected.
    {:time   70 :type :invoke :f :read              :process 0}
    {:time   80 :type :ok     :f :read :value  132  :process 0}

    ;; WHY is reading 'nil' being ignored?  This ought to fail!

    {:time  770 :type :invoke :f :read              :process 0}
    {:time  780 :type :ok     :f :read :value  nil  :process 0}

    ;; This succeeds, as expected.
    {:time 7770 :type :invoke :f :read              :process 0}
    {:time 7780 :type :ok     :f :read :value  132  :process 0}

    ;; If 8770-8780 are included, then model/cas-register fails, as expected.
    ;; :model {:msg "can't read 999 from register 132"}

    ;; {:time 8770 :type :invoke :f :read              :process 0}
    ;; {:time 8780 :type :ok     :f :read :value  999  :process 0}

    ]

The Corfu client that I'd instrumented would assume that
the register's value was nil if the Corfu key (in an SMRMap)
had never been set or if it had been erased.  This led to weird
behavior by Knossos, which would say at that time 780 above that
the register was linearizable when it clearly is not.

This small patch always checks the model's value against the
history.  This patch may not be applicable to Knossos's wider
use in Jepsen, but it caused me more than enough head-scratching
to record it & the why behind it.